+++ src/main/java/app/controller/GameEngine.java
@@ -4,14 +4,11 @@ import app.controller.graphicsEngine.GraphicsEngine;
 import app.controller.linAlg.Intersection;
 import app.controller.linAlg.Vector;
 import app.controller.soundEngine.SoundEngine;
-import app.controller.soundEngine.SoundVector;
 import app.model.Trail;
-import app.model.agents.ACO.AcoAgent;
 import app.model.agents.Agent;
 import app.model.boundary.Boundary;
 import app.model.Map;
 import app.model.boundary.PortalBoundary;
-import app.model.sound.SoundSource;
 import app.view.simulation.Renderer;
 import javafx.animation.Animation;
 import javafx.animation.KeyFrame;
@@ -19,8 +16,6 @@ import javafx.animation.Timeline;
 import javafx.scene.input.KeyEvent;
 import javafx.util.Duration;
 import lombok.Getter;
-
-import java.util.ArrayList;
 
 public class GameEngine
 {
