+++ src/main/java/app/controller/soundEngine/Sound.java
@@ -1,7 +1,7 @@
 package app.controller.soundEngine;
 
 import app.controller.linAlg.Vector;
-import app.model.sound.SoundSource;
+
 import java.util.Objects;
 import lombok.Getter;
