+++ src/main/java/app/controller/soundEngine/SoundSource.java
@@ -1,4 +1,4 @@
-package app.model.sound;
+package app.controller.soundEngine;
 
 import app.controller.linAlg.Intersection;
 import app.controller.linAlg.Vector;
@@ -38,8 +38,7 @@ public class SoundSource
     public ArrayList<SoundVector> heard(Agent agent)
     {
         ArrayList<SoundVector> output = new ArrayList<>();
-        Queue<SoundRay> queue = new ConcurrentLinkedQueue<>();
-        queue.addAll(rays);
+        Queue<SoundRay> queue = new ConcurrentLinkedQueue<>(rays);
 
         while(!queue.isEmpty())
         {
