+++ src/main/java/app/model/boundary/SoundBoundaryFactory.java
@@ -1,4 +1,4 @@
-package app.model.sound;
+package app.model.boundary;
 
 import app.controller.linAlg.Vector;
 import app.controller.settings.SettingsObject;
