+++ src/main/java/app/model/furniture/Furniture.java
@@ -1,7 +1,7 @@
 package app.model.furniture;
 
 import app.model.boundary.Boundary;
-import app.model.sound.SoundBoundary;
+import app.model.boundary.SoundBoundary;
 import app.model.texture.Texture;
 import javafx.scene.canvas.GraphicsContext;
